mvc updated background

+++ LLBACKEND/LL/src/main/java/com/example/Lovelylawnsbe/LL/PlantRepository.java
@@ -2,6 +2,7 @@ package com.example.Lovelylawnsbe.LL;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlantRepository extends JpaRepository<Plant, Long> {
+public interface PlantRepository extends JpaRepository<Plant, Integer> {
 }
+
